+++ src/kard/users/rewards/types/__init__.py
@@ -27,6 +27,8 @@ if typing.TYPE_CHECKING:
     from .eligibility_offer_relationship import EligibilityOfferRelationship
     from .location_attributes import LocationAttributes
     from .location_data import LocationData
+    from .location_partner_id import LocationPartnerId
+    from .location_partner_id_type import LocationPartnerIdType
     from .location_relationships import LocationRelationships
     from .location_sort_options import LocationSortOptions
     from .locations_response_object import LocationsResponseObject
@@ -75,6 +77,8 @@ _dynamic_imports: typing.Dict[str, str] = {
     "EligibilityOfferRelationship": ".eligibility_offer_relationship",
     "LocationAttributes": ".location_attributes",
     "LocationData": ".location_data",
+    "LocationPartnerId": ".location_partner_id",
+    "LocationPartnerIdType": ".location_partner_id_type",
     "LocationRelationships": ".location_relationships",
     "LocationSortOptions": ".location_sort_options",
     "LocationsResponseObject": ".locations_response_object",
@@ -148,6 +152,8 @@ __all__ = [
     "EligibilityOfferRelationship",
     "LocationAttributes",
     "LocationData",
+    "LocationPartnerId",
+    "LocationPartnerIdType",
     "LocationRelationships",
     "LocationSortOptions",
     "LocationsResponseObject",